pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-system"
version = "0.1.0"
description = "Production-ready multi-agent system with LangGraph, LangChain, MCP, MinIO, and ElasticSearch"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Agent System" }]

dependencies = [
    # LangChain / LangGraph ecosystem
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",

    # Langfuse observability + prompt management
    "langfuse>=2.0.0",

    # FastAPI
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",

    # MCP (Model Context Protocol)
    "mcp>=1.0.0",

    # MinIO object storage
    "minio>=7.2.0",

    # ElasticSearch logging — pin to v8 to match ES server 8.x
    "elasticsearch>=8.0.0,<9.0.0",

    # Config & validation
    "pydantic>=2.8.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",

    # HTTP & async
    "httpx>=0.27.0",
    "aiofiles>=24.0.0",

    # Utilities
    "tenacity>=9.0.0",
    "python-multipart>=0.0.9",
    "structlog>=24.0.0",

    # Built-in tool dependencies
    "ddgs>=9.0.0",
    "python-docx>=1.1.2",

    # PostgreSQL — async driver for the connection pool (stores + repositories)
    "asyncpg>=0.29.0",
    # psycopg2 kept for the MCP @modelcontextprotocol/server-postgres child process
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]
adk = [
    "google-adk>=1.28.0",
]

[project.scripts]
agent-system = "agent_system.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "adk: marks tests that run ADK evaluation workflows",
]
